Create The King of Asgard
In the kingdom of Asgard, ODIN, the King has a group of witches. He assigns one witch to each newly captured country (ith country), and is given a wand that helps them communicate with the witch in ⌊i2⌋
the country. The communication is two-way. There is also a super witch tasked with monitoring the flow of information among the witches. 

Recently, the King visited a new country and issued some orders. The witch in that country then spread rumours to all the other witches using her communication. As the super witch, your job is tracking how the rumours and hoaxes spread. You already know which country the King visited.